slides_translator_agent/requirements.txt
a2a-sdk==0.3.11
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4 ; sys_platform == 'darwin'
asttokens==3.0.0
attrs==25.4.0
authlib==1.6.5
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0 ; implementation_name == 'pypy' or platform_python_implementation != 'PyPy'
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6 ; sys_platform == 'win32'
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
distlib==0.4.0
docstring-parser==0.17.0
executing==2.2.1
fastapi==0.121.0
filelock==3.20.0
google-adk==1.18.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.126.1
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.5.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.49.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
identify==2.6.15
idna==3.11
importlib-metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
mako==1.3.10
markupsafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.21.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.3.4
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-google-genai==0.4b0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-genai==0.2b0
packaging==25.0
parso==0.8.5
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
platformdirs==4.5.0
pre-commit==4.3.0
prompt-toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23 ; (implementation_name != 'PyPy' and platform_python_implementation != 'PyPy') or (implementation_name == 'pypy' and platform_python_implementation == 'PyPy')
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.4
rust-just==1.43.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
sqlalchemy-spanner==1.17.1
sqlparse==0.5.3
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
tenacity==9.1.2
tornado==6.5.2
traitlets==5.14.3
ty==0.0.1a25
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2 ; sys_platform == 'win32'
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
